File md5

+++ Realtime/Realtime/Public/Message/LCIMVideoMessage.cs
@@ -58,7 +58,7 @@ namespace LeanCloud.Realtime {
                 Height = h;
             }
             if (File.MetaData.TryGetValue(MessageDataMetaDurationKey, out object duration) &&
-                duration is float d) {
+                duration is double d) {
                 Duration = d;
             }
         }
